Checks for email, organization and location
