+++ app/api/users/core/services.py
@@ -17,6 +17,10 @@ class UserService(APIEntityServiceMixin):
     @property
     def __collection(self):
         return self._EntityServiceMixin__collection
+
+    def get_by_params(self, data):
+        user = self.__collection.get_by_equal_params(data)
+        return user[0] if len(user) != 0 else None
     
     
 class UserInfoService(APIEntityServiceMixin):
@@ -38,7 +42,7 @@ class UserInfoService(APIEntityServiceMixin):
     def get_table_name(cls):
         return cls.__TABLE_NAME
     
-    
+
     def get_by_user_id(self, user_id):
         user_info = self.__collection.get_by_equal_params({'user_id': int(user_id)})
         return user_info[0]
